requirements.txt
urllib3
requests
beautifulsoup4
lxml
Pillow
pytesseract
argparse
